Added refund calculations & most of payment
Refund calculations are done via type enum value conversions.

Payment is mostly implemented, but bugged, the current issue is that
when POST handler tries to access the ticket inner object flight, it is
null, which is strange because the flight object is present on the get
method for pay and within the razor view, why it gets lost I have no
idea.

Posted a SO question:
http://stackoverflow.com/questions/26959377/viewmodel-objects-child-is-null-on-post